requirements/common.txt
amqp==1.4.3
anyjson==0.3.3
Babel==2.1.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.16
bleach==1.4
celery==3.1.9
certifi==2015.11.20.1
click==3.3
contextlib2==0.4.0
cssselect==0.9.1
decorator==4.0.6
Django==1.6.2
django-appconf==0.6
django-celery==3.1.9
django-configurations==0.8
django-debug-toolbar==1.0.1
django-excel-response==1.0
django-extensions==1.5.0
django-grappelli==2.5.2
django-htmlmin==0.7.0
django-pipeline==1.4.6
django-querycount==0.1.1
django-uuslug==1.0.2
eventlet==0.17.4
Flask==0.10.1
Flask-Cache==0.13.1
flower==0.8.3
funcsigs==0.4
functools32==3.2.3.post2
futures==2.2.0
gitdb==0.6.4
GitPython==0.3.6
greenlet==0.4.9
gunicorn==19.1.0
html5lib==0.999
icalendar==3.6.2
ipython==1.2.1
itsdangerous==0.24
jieba==0.35
Jinja2==2.8
jsonpath-rw==1.4.0
jsonschema==2.5.1
kombu==3.0.12
lxml==3.4.1
manifesto==0.5
MarkupSafe==0.23
mock==1.3.0
MySQL-python==1.2.5
netaddr==0.7.18
PasteDeploy==1.5.2
pbr==1.8.1
pexpect==4.0.1
Pillow==2.5.1
pingpp==2.0.0
ply==3.8
ptyprocess==0.5
pyExcelerator==0.6.4.1
pyodbc==3.0.6
pyteaser==1.0
python-dateutil==2.2
python-slugify==0.0.7
pytz==2013.9
pyvmomi==6.0.0
PyYAML==3.11
qiniu==7.0.6
raven==4.0.4
readline==6.2.4.1
redis==2.10.5
requests==2.9.1
rq==0.5.6
rq-scheduler==0.5.1
singledispatch==3.4.0.3
six==1.10.0
smmap==0.9.0
South==0.8.4
SQLAlchemy==0.8.2
sqlparse==0.1.11
stevedore==1.10.0
suds==0.4
tornado==4.3
Unidecode==0.4.14
upyun==2.1.1
vcrpy==1.7.4
wechat-sdk==0.5.9
Werkzeug==0.11.3
wheel==0.24.0
wrapt==1.10.5
xlrd==0.9.2
xlutils==1.7.1
xlwt==0.7.5
xmltodict==0.9.0
ewsclient==0.3.1
